Add AsyncAdapter type
AsyncAdapter can be used to handle Line, Options, and Command events in
an asynchronous manner, similar to the main Yarn Spinner implementation.
AsyncAdapter is a DialogueHandler that can be passed to VirtualMachine.
Then, the user can supply an AsyncDialogueHandler. When Line, Options,
or Command is called on the AsyncDialogueHandler, the VM execution is
paused, and can continue by calling AsyncAdapter.Go, GoWithChoice,
or can be stopped entirely with AsyncAdapter.Abort.